store video file named by session id

+++ src/editor/editor-server.ts
@@ -1,10 +1,10 @@
 import * as http from 'http'
-import fs from 'fs'
+import fs, { WriteStream } from 'fs'
 import path from 'path'
 import { WebSocketServer } from 'ws'
 import { Dependencies, EndpointInit, WSInit } from '../server'
 import { consoleLogger, logPrefix } from '../logger'
-import { ChangeRecord } from './editor-types'
+import { ChangeMessage } from './editor-types'
 import { msgPayload } from '../websocket'
 import { invariant } from '../errors'
 
@@ -55,10 +55,9 @@ export const initWS: WSInit = ({sql}: Dependencies, server: http.Server) => {
       }
 
       if (msgStr === 'start') {
-        logger.info('starting session')
         const id = await sql<{id: number}[]>`insert into sessions (changes) values ('[]') returning id`
         sessionId = id[0].id
-        logger.info('sending session', sessionId)
+        logger.info('starting session', sessionId)
         ws.send(`start ${sessionId}`)
         return
       }
@@ -81,16 +80,26 @@ export const initWS: WSInit = ({sql}: Dependencies, server: http.Server) => {
   })
 
   videoWSS.on('connection', (ws) => {
-    const fileName = `${tmpDir}/video.webm`
-    fs.mkdirSync(path.dirname(fileName), { recursive: true });
-    try {
-      fs.rmSync(fileName)
-    } catch {}
-
-    const fileStream = fs.createWriteStream(`${tmpDir}/video.webm`, { flags: 'a' });
-    logger.info(`starting to write video stream to file: ${fileName}`)
-
+    let firstMessage = true
+    let fileStream: WriteStream
     ws.on('message', (message) => {
+      if (firstMessage) {
+        firstMessage = false
+        const strMessage = message.toString()
+        invariant(strMessage.startsWith('start'), 'first message in video stream must be start message')
+        const sessionId = msgPayload('start')(strMessage)
+
+        const fileName = `${tmpDir}/${sessionId}.webm`
+        fs.mkdirSync(path.dirname(fileName), { recursive: true });
+        try {
+          fs.rmSync(fileName)
+        } catch {}
+
+        fileStream = fs.createWriteStream(fileName, { flags: 'a' });
+        logger.info(`starting to write video stream to file: ${fileName}`)
+        return
+      }
+
       // Only raw blob data can be sent
       fileStream.write(Buffer.from(new Uint8Array(message as any)));
     })
